+++ builds/v0.3/release_notes.md
@@ -8,3 +8,4 @@
 * Fixed header alignment on tasks view
 * Locked editing rewards box when "Filler" is selected
 * Fixed scrolling in the tasks section... again.
+* Fixed a bug where DeathLink tag wasn't being sent on reconnection